Homework 1
Still, bugs exist. Tried to make it work though. Original intention:
detects the letter of the txt file one by one, and change the font
color accordingly. E.g: if the letter is b, changes it to blue, if it's
r, changes it to red. however doesn't work all the way along.